// src/typings/user.enum.ts
export enum FreeCardsTeamStatusEnum {
  NONE = 'none',
  SKIP = 'skip',
  DONE = 'done',
}

export enum Jersey {
  DEFAULT1 = 'default1',
  DEFAULT2 = 'default2',
  DEFAULT3 = 'default3',
}

export enum Sponsor {
  CYLIMIT_WHITE = 'cylimit_white',
  CYLIMIT_BLACK = 'cylimit_black',
  CYLIMIT_RED = 'cylimit_red',
  CYLIMIT_YELLOW = 'cylimit_yellow',
  CYLIMIT_BLUE = 'cylimit_blue',
}

export enum KycStatus {
  NEW = 'new',
  PENDING = 'pending',
  REJECTED = 'rejected',
  APPROVED = 'approved',
}